requirements.txt
streamlit==1.35.0
python-dotenv==1.0.1
groq==0.30.0
PyMuPDF
streamlit-extras
firebase-admin
pyrebase4
pycryptodome
setuptools
google-cloud-storage
reportlab
python-docx
pdf2docx
language_tool_python